Add support for automatic reporting to a bug tracker
This PR adds support for automatic status reports for batches that are submitted by workers. All batch entries are listed in the ticket. If all entries were successful the ticket is automatically closed. If any failed, then the ticket remain opens. Rescuing an entry will automatically update the ticket. When all entries are rescued, the ticket will be closed by the bot.